Migration updated

+++ database/migrations/2018_08_11_110958_add_currency_id_to_suppliers_table.php
@@ -25,8 +25,12 @@ class AddCurrencyIdToSuppliersTable extends Migration
      */
     public function down()
     {
-        Schema::table('suppliers', function (Blueprint $table) {
-            $table->dropColumn('currency_id');
-        });
+        if (Schema::hasColumn('suppliers', 'currency_id'))
+        {
+            Schema::table('suppliers', function (Blueprint $table)
+            {
+                $table->dropColumn('currency_id');
+            });
+        }
     }
 }
